fix : 이미시 삭제 함수 수정

+++ src/main/java/com/bonheur/domain/image/service/ImageServiceImpl.java
@@ -47,15 +47,16 @@ public class ImageServiceImpl implements ImageService{
         uploadImages(board, images);            //새로운 이미지 업로드
     }
 
-    // # image s3에서 삭제
+    // # image s3, 테이블에서 삭제
     @Override
     @Transactional
-    public void deleteImages(Board board) {
+    public void deleteImagesIns3(Board board) {
         List<Image> toDeleteImages = imageRepository.findAllByBoard_Id(board.getId());
 
-        for (Image img : toDeleteImages) {
-            fileUploadUtil.deleteFile(img.getPath());
-            imageRepository.delete(img);
+        if (toDeleteImages != null) {
+            for (Image img : toDeleteImages) {
+                fileUploadUtil.deleteFile(img.getPath());
+            }
         }
     }
 }
